requirements.txt
customtkinter==5.2.2
darkdetect==0.8.0
packaging==23.2
pillow==10.2.0
six==1.16.0
